# [Enhancement] Optimize parallel merge algorithm