## Why I'm doing:

For some special data distribution, the parallel merge algorithm has really bad performance.

## What I'm doing:

Improve the performance of the corner cases.

## Test

```sql
DROP DATABASE IF EXISTS `analytic`;
CREATE DATABASE `analytic`;
USE `analytic`;
DROP TABLE IF EXISTS `t0`;
CREATE TABLE IF NOT EXISTS `t0` (
  `id_int` INT(11) NOT NULL,
  `id_tinyint` TINYINT NOT NULL,
  `id_smallint` SMALLINT NOT NULL,
  `id_bigint` BIGINT NOT NULL,
  `id_largeint` LARGEINT NOT NULL,
  `id_float` FLOAT NOT NULL,
  `id_double` DOUBLE NOT NULL,
  `id_char` CHAR(10) NOT NULL,
  `id_varchar` VARCHAR(100) NOT NULL,
  `id_date` DATE NOT NULL,
  `id_datetime` DATETIME NOT NULL,
  `id_decimal` DECIMAL(9,0) NOT NULL
) ENGINE=OLAP
DUPLICATE KEY(`id_int`)
DISTRIBUTED BY HASH(`id_int`) BUCKETS 10
PROPERTIES (
 "replication_num" = "1"
);

INSERT INTO `t0` (id_int, id_tinyint, id_smallint, id_bigint, id_largeint, id_float, id_double, id_char, id_varchar, id_date, id_datetime, id_decimal) VALUES
    (1, 10, 100, 1000, 10000, 100000.1, 1000000.1, 'some_char', 'some_varchar', '2020-07-24', '2020-07-24', 123.78),
    (1, 20, 200, 2000, 20000, 200000.1, 2000000.1, 'some_char2', 'some_varchar2', '2020-07-23', '1020-07-24', 123.56),
    (1, 30, 300, 3000, 30000, 300000.1, 3000000.1, 'some_char3', 'some_varchar3', '2020-07-22', '3020-07-24', 123.34),
    (2, 40, 400, 4000, 40000, 400000.1, 4000000.1, 'some_char4', 'some_varchar4', '2020-07-21', '1999-07-24', 123.12),
    (2, 50, 100, 5000, 10000, 100000.1, 1000000.1, 'some_char', 'some_varchar', '2020-07-24', '2020-07-24', 1234.11),
    (3, 60, 200, 6000, 20000, 200000.1, 2000000.1, 'some_char2', 'some_varchar2', '2020-07-23', '1020-07-24', 1234.12);

INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;
INSERT INTO `t0` SELECT * FROM `t0`;

-- Now the t0 has 200'000'000 rows, with 50% id_int=1, 33% id_int=2, 17% id_int=3
-- And we here to insert more values of id_int in the range(0, 1000000) of around 20'000'000 rows.
-- So the main skew data will holds 90% of rows, with others take the remain part.
INSERT INTO `t0` select cast(rand() * 1000000 as int), id_tinyint, id_smallint, id_bigint, id_largeint, id_float, id_double, id_char, id_varchar, id_date, id_datetime, id_decimal from `t0` limit 20000000;
```

```sql
-- Q1
SELECT SUM(id_tinyint), SUM(id_smallint), SUM(id_bigint), SUM(id_largeint), SUM(id_float), SUM(id_double), MIN(id_char), MIN(id_varchar), MIN(id_date), MIN(id_datetime), SUM(id_decimal),
       SUM(lag_smallint), SUM(lag_bigint), SUM(lag_largeint), SUM(lag_float), SUM(lag_double), MIN(lag_char), MIN(lag_varchar), MIN(lag_date), MIN(lag_datetime), SUM(lag_decimal)
 FROM (
    SELECT *,
        LAG(id_smallint, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_smallint,
        LAG(id_bigint, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_bigint,
        LAG(id_largeint, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_largeint,
        LAG(id_float, 1) OVER ([skewed]PARTITION BY id_int ORDER BY id_tinyint) AS lag_float,
        LAG(id_double, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_double,
        LAG(id_char, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_char,
        LAG(id_varchar, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_varchar,
        LAG(id_date, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_date,
        LAG(id_datetime, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_datetime,
        LAG(id_decimal, 1) OVER (PARTITION BY id_int ORDER BY id_tinyint) AS lag_decimal
    FROM `t0` WHERE id_int = 1
) tmp2;
```

| Query | Before | After |
|:--|:--|:--|
| Q1 | 797s | 55s |

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5