@stdpain 2024-06-17
https://github.com/Mergifyio backport branch-3.2